Rename matplotlib compat module to utils
Replaces `deeplabcut.utils.matplotlib_compat` with `matplotlib_utils` and narrows it to the shared `silence_axes_logger` helper plus `AXES_LOGGER_NAME`. Removes the old colormap and artist-removal compatibility helpers and rewrites tests to target the new module and its logger behavior.
